+++ NEWS.md
@@ -1,6 +1,7 @@
 * Add libraries script to templates
 * Add knitr options script to templates
 * Add `clamour_load_tweets()` function
+* Add buttons to hide parameters and session info
 
 # clamour 0.0.0.9011 (2020-09-01)