+++ NEWS.md
@@ -2,6 +2,7 @@
 * Add knitr options script to templates
 * Add `clamour_load_tweets()` function
 * Add buttons to hide parameters and session info
+* Improve formatting of topics tweets tables
 
 # clamour 0.0.0.9011 (2020-09-01)